requirements.txt
# Web 框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9  # PostgreSQL
aiosqlite==0.19.0       # SQLite 异步支持
asyncpg==0.29.0

# 缓存
redis[hiredis]==5.0.1
aioredis==2.0.1

# 任务调度
apscheduler==3.10.4

# HTTP 客户端
httpx==0.25.2
aiohttp==3.9.1

# AI 服务
openai==1.3.7
anthropic==0.7.8

# 配置管理
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
email-validator==2.1.0

# YAML 配置
PyYAML==6.0.1

# 邮件发送
aiosmtplib==3.0.1
jinja2==3.1.2

# 日志
loguru==0.7.2

# 工具类
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7

# 开发和测试
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# 安全
cryptography==41.0.8
bcrypt==4.1.2
python-jose[cryptography]==3.3.0

# 数据验证
marshmallow==3.20.1

# 数据处理
pandas==2.1.4
numpy==1.25.2

# 模板引擎
jinja2==3.1.2

# 文件操作
aiofiles==23.2.1

# JSON Web Tokens
pyjwt[crypto]==2.8.0

# 开发和测试
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# 监控
prometheus-client==0.19.0

# 预提交
pre-commit==3.6.0 